requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cupy-cuda102==9.0.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
detectron2==0.4+cu102
fastrlock==0.6
future==0.18.2
fvcore==0.1.3.post20210317
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
iopath==0.1.8
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.2.52
Pillow==8.2.0
portalocker==2.3.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.4.8
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
yacs==0.1.8
zipp==3.4.1
